EES-2687 - broke out MethodologyApprovalService from MethodologyService
