+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/Interfaces/Methodologies/IMethodologyApprovalService.cs
@@ -1,0 +1,19 @@
+﻿#nullable enable
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using GovUk.Education.ExploreEducationStatistics.Admin.ViewModels;
+using GovUk.Education.ExploreEducationStatistics.Admin.ViewModels.Methodology;
+using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using GovUk.Education.ExploreEducationStatistics.Content.Model;
+using Microsoft.AspNetCore.Mvc;
+
+namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces.Methodologies
+{
+    public interface IMethodologyApprovalService
+    {
+        Task<Either<ActionResult, MethodologyVersion>> UpdateApprovalStatus(
+            Guid methodologyVersionId,
+            MethodologyApprovalUpdateRequest request);
+    }
+}
