+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/MethodologyVersion.cs
@@ -127,5 +127,15 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model
 
             return copy;
         }
+
+        public void SetToDraft()
+        {
+            PublishingStrategy = Immediately;
+            Status = Draft;
+            ScheduledWithRelease = null;
+            ScheduledWithReleaseId = null;
+            InternalReleaseNote = null;
+            Updated = DateTime.UtcNow;
+        }
     }
 }
